Update __init__.py
I thought it is a bit messy, that the Modal does only half of the state management.
Big change: 
Modal accepts a callable containing the content which is rendered if modal is open. Now modal can handle state, no need for a "if modal.is_open()" outside of the modal

Small change:
I replaced the button-true-false-state by a callback. This can be passed to any button. You could still re-add the open and false methods. I also think this is as it stands a breaking change. Yet I think it still should be possible to have backward compatibility by adding back the yield statement of the container. 
I guess this is all easily done, if people think those changes are a good idea. The main magic is the css bit.